backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==10.4.0
django-filter==24.1
djangorestframework-simplejwt==5.3.1
sendgrid==6.12.4
django-sendgrid-v5==1.3.0
